pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-quant-research-agent"
version = "0.1.0"
description = "LLM-planned executable workflows for quant research experiments."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "openai>=1.0.0",
  "baostock>=0.8.9",
]

[project.scripts]
quant-agent = "quant_research_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

